requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
python-multipart==0.0.9

qdrant-client>=1.12.0

numpy<2.0.0
torch
transformers==4.44.2
huggingface-hub==0.25.2
sentence-transformers>=2.7.0
pylate>=1.0.0
openai>=1.30.0
httpx==0.27.2
tiktoken==0.6.0
scikit-learn==1.4.1.post1
accelerate>=0.31.0

python-docx==1.1.0
psutil==5.9.8
langchain==0.1.0